Project 4.6 | Test

+++ src/main/java/edu/project4/Pixel.java
@@ -2,7 +2,7 @@ package edu.project4;
 
 import java.util.concurrent.ThreadLocalRandom;
 
-public record Pixel(int r, int g, int b, double hitCount) {
+public record Pixel(double r, double g, double b, double hitCount) {
 
     @SuppressWarnings("MagicNumber")
     public static int randomColor() {
